Move fc_tools from util to time_series/spectre
- update test imports
